+++ src/core/constants.ts
@@ -10,10 +10,10 @@ import { Separator, Config, NameIndex } from '../models/index';
  * The current version of `Namefully`.
  * @constant
  */
-export const version = '1.0.8';
+export const version = '1.0.9';
 
 /**
- * CONFIG constant
+ * CONFIG
  * @constant
  * @type {Config}
  * @default
@@ -25,10 +25,10 @@ export const CONFIG: Config = {
     ending: Separator.SPACE,
     bypass: false,
     lastnameFormat: 'father',
-}
+} as const
 
 /**
- * NAME_INDEX constant
+ * NAME_INDEX
  * @constant
  * @type {NameIndex}
  * @default
@@ -39,4 +39,59 @@ export const NAME_INDEX: NameIndex = {
     middlename: 2,
     lastname: 3,
     suffix: 4,
+} as const
+
+/**
+ * RESTRICTED_CHARS
+ * @constant
+ * @default
+ */
+export const RESTRICTED_CHARS = [
+    Separator.SPACE,
+    Separator.SINGLE_QUOTE,
+    Separator.HYPHEN
+] as const
+
+
+class CharSet<T> extends Set<T> {
+    random(): T {
+        return Array.from(this)[Math.floor(Math.random() * this.size)];
+    }
 }
+
+/**
+ * PASSWORD_MAPPER
+ * @constant
+ */
+export const PASSWORD_MAPPER = new Map([
+    [ 'a', new CharSet(['a', 'A', '@', '4']) ],
+    [ 'b', new CharSet(['b', 'B', '6', '|)', '|3', '|>']) ],
+    [ 'c', new CharSet(['c', 'C', '(', '<']) ],
+    [ 'd', new CharSet(['d', 'D', '(|', '<|']) ],
+    [ 'e', new CharSet(['e', 'E', '3']) ],
+    [ 'f', new CharSet(['f', 'F', '7']) ],
+    [ 'g', new CharSet(['g', 'G', '8', '&', '**']) ],
+    [ 'h', new CharSet(['h', 'H', '#', '|-|']) ],
+    [ 'i', new CharSet(['i', 'I', '!', '1', '|', '--']) ],
+    [ 'j', new CharSet(['j', 'J', '|-']) ],
+    [ 'k', new CharSet(['k', 'K', '%', '*|*']) ],
+    [ 'l', new CharSet(['l', 'L', '1', '!', '|_']) ],
+    [ 'm', new CharSet(['m', 'M', '[|]']) ],
+    [ 'n', new CharSet(['n', 'N', '|-|', '!=', '++']) ],
+    [ 'o', new CharSet(['o', 'O', '0', '*']) ],
+    [ 'p', new CharSet(['p', 'P', '|3', '|)']) ],
+    [ 'q', new CharSet(['q', 'Q', '&']) ],
+    [ 'r', new CharSet(['r', 'R', '7', '[']) ],
+    [ 's', new CharSet(['s', 'S', '5', '$']) ],
+    [ 't', new CharSet(['t', 'T', '7']) ],
+    [ 'u', new CharSet(['u', 'U', '|_|']) ],
+    [ 'v', new CharSet(['v', 'V', '>', '<']) ],
+    [ 'w', new CharSet(['w', 'W', '[|]']) ],
+    [ 'x', new CharSet(['x', 'X', '%', '#']) ],
+    [ 'z', new CharSet(['z', 'Z', '2', '!=']) ],
+    [ '$', new CharSet([
+        '!', '@', '#', '$', '%', '^', '&', '*', '(', ')', '-', '+', '[',
+        '_', '=', '{', '}', ':', ';', ',', '.', '<', '>', '|', '~', ']',
+        '0', '1', '2', '3', '4', '5', '6', '7', '8', '9'
+    ])],
+]);