+++ src/validators/suffix.validator.ts
@@ -19,7 +19,7 @@ export default class SuffixValidator implements Validator<string> {
      * @param {string} value data to validate
      */
     validate(value: string): void {
-        const suffixes: Array<string> = Object.entries(Suffix).map(e => e[1].toLowerCase()); // values
+        const suffixes: string[] = Object.entries(Suffix).map(e => e[1].toLowerCase()); // values
         if (suffixes.indexOf(value.toLowerCase()) === -1)
             throw new ValidationError(`unknown value '${value}'`, 'Suffix');
     }